requirements.txt
numpy==1.21.6
scipy==1.6.2
astropy==4.0.2
matplotlib==3.3.1
requests==2.24.0
python-dotenv==0.14.0
cmasher==1.6.0
Cython==3.0.2
photutils
pandas
opencv